- Updated serverRequest.js to throw error in the catch promise - Created apiURL.js file to save url, to reduce repetition - Moved apiServer directory into server folder - Created helpFunctions/helperFunction.js to store all of the server helper functions
