+++ server/apiServer/apiRequest.js
@@ -1,12 +1,14 @@
 const axios = require('axios');
-const { apiKey } = require('../configure.js');
+const { apiKey } = require('../../configure.js');
+const { url } = require('../apiURL.js');
 
 axios.defaults.headers.common['Authorization'] = apiKey;
 module.exports = {
 
 // GET REQUEST TO THE Atelier Products API
-  get: (url, callback) => {
-    axios.get(url)
+  get: (endpoint, callback) => {
+    //console.log(url.endpoint);
+    axios.get(url+ endpoint)
       .then((result) => {
         callback(null, result);
       })
