+++ server/apiURL.js
@@ -1,0 +1,3 @@
+const config = require('../configure.js');
+
+module.exports = { url: `https://app-hrsei-api.herokuapp.com/api/fec2/${config.CAMPUS_CODE}`}
